Runtime linking with GTK on Linux
