+++ GNUmakefile
@@ -164,8 +164,7 @@ DEFS := $(DEFS) \
 
 INCLUDES := $(INCLUDES) \
 	-Isrc/unix/linux \
-	-Isrc/unix/linux/x11 \
-	$(sort $(shell pkg-config --cflags x11 gtk+-3.0 webkit2gtk-4.0))
+	-Isrc/unix/linux/x11
 
 endif
