+++ src/gfx/vk/vkproc.h
@@ -19,6 +19,7 @@
 		((e) == VK_ERROR_OUT_OF_DATE_KHR || (e) == VK_SUBOPTIMAL_KHR)
 
 #elif defined(MTY_VK_XLIB)
+	#define LIBX11_NO_SYMBOLS
 	#include "dl/libx11.h"
 	#include "vulkan/vulkan_xlib.h"
